Refactor(view): merged chatBoard into viewGroup
